saved changes

+++ src/main/java/com/cpiura/catics/repository/GlpiNewTrabajadoresTorreRepository.java
@@ -1,11 +1,10 @@
 package com.cpiura.catics.repository;
 
+import com.cpiura.catics.entity.GlpiNewTrabajadoresTorre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.cpiura.catics.entity.GlpiNewTrabajadoresTorre;
-
 @Repository
-public interface GlpiNewTrabajadoresTorreRepository extends JpaRepository<GlpiNewTrabajadoresTorre, Integer> {
-    // Métodos personalizados si es necesario
+public interface GlpiNewTrabajadoresTorreRepository extends JpaRepository<GlpiNewTrabajadoresTorre, Long> {
+    // Puedes agregar métodos personalizados si lo necesitas
 }
